Fix ParseStream interface to allow extending parent nodes leftward
When parsing infix constructs, child nodes are discovered before their
parents so there's a need to reparent nodes which have already been
emitted into the event stream. This change adds emit(::ParseStream),
allowing a leftmost byte position from position(::ParseStream) to be
saved and passed to that to emit any number of parent nodes
retroactively.

This is a bit like rust-analyzer's event stream, but recording the
current position keeps the state in the parser rather than requiring
Start markers in the event stream. There's no need to abort a node -
simply don't emit() it and no chasing a linked list of forward_parent
indices. (Instead, we need to iterate backward through the text spans to
find children for each node when building the tree.)