Split out green tree and make it language independent
Language independence of the green tree allows us to use the low-level
data structure to overlay the text of any formal language. For example,
it would be neat to have an s-expression parser :-)

It also allows for neater experimentation with the representation of the
GreenNode's head.  For example, do we want a 16-bit Kind? Do we want
flags in there? How about memorizing the layout of the child nodes which
the parser knew about?