Consistent error reporting for raw->AST interface
Errors are now reported by emitting a node of kind K"error" in the
appropriate location and arranged so that the transformation from green
tree to AST for a given node kind is always deterministic. This provides
a much clearer contract between the parser and the downstream consumers
of the green tree.

Error nodes can encompass zero or several source code tokens, and be
marked as trivia if they need to be ignored as part of the AST
transformation.