Change `shm_open()` calling ABI on aarch64 Darwin
Beacuse `shm_open()` is a variadic function, if we don't declare it as
such, the kernel receives trash as the `permissions` value, which
occasionally results in errors when creating shared memory segments.
This did not happen on `x86_64` because the calling convention doesn't
change so much between variadic and non-vadiadic functions.