Mark stack as non-executable
The linker is unable to detect that executable stack is not required
and so marks libjulia.so as requiring it
Pass `-Wl,-z,noexecstack` to ensure that the stack is marked as not executable.
This improves security and allows Julia to run on systems where SELinux has been
configured to disallow executable stacks.

AFAIK no change is needed on OSes other than Linux as they do not allow executable stacks
by default.