optimizer: improve SROA performance
Essentially this commit reduces the runtime count # of domtree construction.
The key observation is that SROA of mutable allocations isn't often successful.
More specifically, during the sysimg build, `sroa_mutables!` is called
`5776` times, and actual SROA happened only for `126` cases out of those cases.
(, which means we should be able to avoid domtree construction in `5650`/`5776` cases).