Rewrite how macro names are handled + parse export
Macro names are now represented with a special token kind so they can be
locally mapped into Julia Exprs rather than having the `@` added by
using the context of the containing expression. Modify parse_call_chain
to use this, and make use of it to implement parsing of export lists.