Don't set HAVE_SSP unconditionally on aarch64
There is no plausible reason for this. And indeed, this patch fixes
compilation for the libjulia_jll builder on Yggdrasil which currently fails
due to the absence of libssp.

The removed line was introduced in PR #41936 but without any justification
that I could discern, so it might have just slipped in there accidentally.