win32: Fix SEH frame sentinel (#43570)
The last entry sentinel for Win64 SEH is `~0L` not NULL. Apparently
this doesn't cause issues on windows proper, but does crash wine.
Arguably if Windows doesn't have issues then we should just fix
this in wine, but since we control the source and nobody else
ever seems to have run into this, let's just fix it and save
the good Wine folks some headache. Fixes #43569.