Fix SnoopCompile's `snoopl` macro and add test. (#43634)
* Fix SnoopCompile's `snoopl` macro and add test.

Previously the `@snoopl` functionality from SnoopCompile wasn't unit
tested at all, and so it was broken in https://github.com/JuliaLang/julia/pull/38160,
which changed the requirements for exporting C functions from the
julia-internal shared lib.

This commit restores the functionality by exporting it correctly, and
also adds a unit test for snoopl, to make sure it isn't broken in the
future. Sorry that we didn't test it in the first place! :)