Add `get_preferences()` to load all prefs in the current environment
Previously, we had the machinery to load per-UUID, but some tooling
needs to be able to load all preferences (without loading
`Preferences.jl`), so it's better to support being able to load
everything from just within Base.

Also add some preference loading tests, so that we can fearlessly
refactor this code in the future.