Tokenize: emit all strings as STRING or CMD
Emit strings as either STRING or CMD. This distinction is required
downstream from the lexer when processing the escape sequences in raw
strings. (Surprisingly no other cases need this.)

Remove TRIPLE_STRING and TRIPLE_CMD, as that information is now
available in the delimiters which are emitted as separate tokens.

Remove the triplestr flag from the token structs, as that turned out not
to be very useful to the parser.