Make roots fully relocatable
This abandons the attempt to put everything into a single index
and instead sets up a table of blocks indexed by the build_id of the
toplevel module.

Most of the machinery is hidden behind an API defined in methods.c,
so that one could modify the implementation fairly simply.

Small fixes