Clean up early prototype of syntax interpolation
Now that we have a real parser, we can just use this parser for the
syntax interpolation prototype rather than constructing the syntax tree
by hand. It's much nicer!

Also remove the test/syntax_trees.jl prototype. All this is now
implemented so the prototype doesn't seem that useful anymore.