Fix splatnew test
The function being tested currently throws:
```
julia> _construct_structwithsplatnew()
ERROR: TypeError: in new, expected Int64, got a value of type String
Stacktrace:
 [1] StructWithSplatNew
   @ ./REPL[15]:3 [inlined]
 [2] _construct_structwithsplatnew()
   @ Main ./REPL[16]:1
 [3] top-level scope
   @ REPL[18]:1
```

As a result, compiler precision improvements can cause the test
to fail, which is not intended. Fix the fieldtype of the struct
to make sure it doesn't throw.