Add fallback `parse` method that uses `tryparse`
Malformed strings must be expected when parsing. In that case, `parse`
throws an error, which means handling it neccesitates catching, which
is slow an un-idiomatic in Julia. For this reason, `tryparse` was added.

Because `parse` can generically be derived from `tryparse`, if we implement
`parse` in terms of `tryparse`, then implementing the latter give both methods
for free, such that the caller has control over how to handle failures.

This PR adds a fallback definition of `parse` which calls `tryparse`, and
adds a generic docstring to both functions that explain that users should
implement `tryparse` for custom types.

Previously, the `parse` documentation in both docstring and manual implied
that only numbers were being parsed, e.g. `Base.parse` was listed under
"Numbers" in the manual, and the generic docstring referred to numbers.
This PR moves the manual entry to "Strings" and clarifies that the old
docstring refers to parsing numbers only.

See also: #43149