In `Base.runtests`, after tests fail, load `InteractiveUtils` from `@stdlib` (#43804)
In `Base.runtests`, if the tests fail, before loading `InteractiveUtils`, set the `LOAD_PATH` to only include `@stdlib`.

In theory, it's possible that someone has a setup where their load path has a package named `InteractiveUtils`. So, to increase the probability that we are actually loading the correct `InteractiveUtils` stdlib, let's set `Base.LOAD_PATH` to only include `@stdlib`.

After we load `InteractiveUtils`, we restore the original contents of `Base.LOAD_PATH`.