deleteat!: Handle unassigned array elements (#43941)
This function was erroring on unassigned array elements. This is
the minimal fix to resolve that issue. I tried something more
generic, but it turns out all of our Array code basically assumes
that all elements are assigned. Even basic things like `==`, `hash`,
etc. error for arrays with unassigned elements. I think we may want
to strongly consider removing the ability to unassign elements in
arrays in favor of Union{Nothing, T} arrays, which our existing code
handles fine and should be able to be made to have equivalent performance.