Make parser API input and ParseStream text input coherent
Internally in the parser we'd like to work with a string of UTF-8
encoded source code, as we'd like random access. We'd also not like to
copy the text before parsing it. However some objects like IOBuffer
work with a buffer of bytes rather than a string, and a string can't be
constructed without copying the bytes.

The best common, concrete and no-copy representation seems to be
Vector{UInt8}, so settle on this for the concrete representation of
code. (This is also easy to pass to the current Lexer implementation.)

With this convention, make the ParseStream, parse and parseall functions
accept any of a family of types, all of which won't be copied:
- (sub)strings and buffers
- IOs
- Raw ptr,len

textbuf::Vector{UInt8}
text_root::Any