llvm-cpufeatures: get TargetMachine from the MachineModuleInfoWrapperPass pass
This is usually only supposed to be accessible to a MachineModulePass,
but we can trick llvm to give us access to the TargetMachine here too.

Also hack in a sysimg check also, so that the compile result is ensured
to be compatible with the loaded image too (we disable loading of the
sysimg when emitting a new compile).