Make inbounds taint consistency
Inbounds and `--check-bounds=no` basically make the assertion:
If this is dynamically reached during exceution then the index
will be inbounds. However, effects on a function are a stronger
statement. In particular, for *any* input values (not just the
dynamically reached ones), the effects need to hold. This is
in particular true, because inference can run functions that
are dynamically dead, e.g.

```
if unknown_bool_return() # false at runtime, but inference doesn't know
    x = sin(1.0)
end
```

Inference will (and we want it to) run the `sin(1.0)` even
though it is not dynamically reached.

For the moment, make any use of `--check-bounds=no` or `@inbounds`
taint the consistency effect, which is semantically meaningful and
prevents inference from running the function. In the future, we
may want more precise tracking of inbounds that would let us
recover some precision here.