doc: add intuitive rules on broadcasting
Broadcasting is a very common trick in scientific computation, but
Julia manual itself does not include the semantic of broadcasting.

I notice there are many new students and researchers who are unfamilar with the
"rules", I learned it from the numpy manual
https://numpy.org/doc/stable/user/basics.broadcasting.html so I figure it's
also a good addition to have our Julia version.