Add specialized sorting for Bool
Sorting used to be okay for booleans, but radix sort messed with the
dispatch and made it up to 6 times slower in the worst edge case. Rather
than deal with that dispatch, I added compile time dispatch to a
specialized boolean sorting routine that works on arbitrary orderings
and is twice as fast as what was previously in base.