Speed up fieldtype_tfunc for Union fields
Previously, `fieldtype_tfunc` would return Union fields as
`Type{Union{...}}` rather than `Const(Union{...})`. However,
converting the latter to the former actually takes a fair
amount of time, so try to be a bit more precise here and
check whether the lattice element we're calling fieldtype
on is guaranteed to produce a `Const` argument.