Avoid O(N^2) behavior in widenreturn
In functions with lots of slots that construct PartialStructs with
lots of fields, `widenreturn` would attempt to repeatedly find an
interesting slot for Interconditional to use. This is unnecessary,
because the slot selected does not depend on the value of the
PartialStruct field. Memoize this computation and only do it once
per return statement.