Add comments to radaix_sort! and make computationally negligable changes to help the code match the explanation
radix_sort! is not exported and only called with Int lo and hi values, 
so changing the eltype of counts to UInt should have no exported impact. 
I made this change 
a) to simplify slightly
b) because unsigned fails for BitInt
c) because line 715 can now overflow as of the previous commit, and if 
we are using a custom integer type, that might error.

unsigned(hi-lo) gives a false illusion of generality.