whitespace check: rewrite in Julia, add checks
This now checks for:

- trailing non-ASCII whitespace
- non-breaking spaces anywhere
- non-UNIX line endings
- trailing blank lines
- no trailing newline

Git can't handle this, largely because whether it interprets files as
UTF-8 or Latin-1 depends on how system libraries that it uses for regex
matching are compiled, which is inconsistent and hard to fix. The end of
file checks are also quite awkard and inefficient to implement with Git
and shell scripting. Julia is fast and lets us present results clearly.