Add indirection in == fallback
The fallback for `==` just calls `===`. However, we also have types
like `WeakRef` that add three specializations of `==` that just
unwrap the `WeakRef`. It turns out this pattern is also used in
packages, and one of them, ChainRulesCore, has >2300 dependent
packages as of now. Unfortunately, when ChainRulesCore is loaded,
it invalidates code that calls `==` on `WeakRef`. This includes
much of the Serialization stdlib.

This addresses the problem by defining the notion of "unwrapping
for ==" by introducing the non-exported function `unwrap_isequal`.
The fallback is just

```
unwrap_isequal(x) = x
```

but one can add specializations.