[LibGit2] Teach tests to be resilient to `init.defaultBranch` (#44629) (#44745)
If a user runs the tests with a `~/.gitconfig` that provides an
`init.defaultBranch` that is not `master`, our tests fail.  Let's adjust
to the user's configuration as appropriate.  We'll also rename this to
`default_branch` to signify that it may not be called `master` anymore.

(cherry picked from commit ac1d69302ce8e682bef3108296f5a6c182b62e5f)

Co-authored-by: Elliot Saba <staticfloat@gmail.com>