Make multiversioning bugpoint robust
Bugpoint can introduce IR forms that would cause multiversioning to
fail and that julia should never emit. However, if we actually crash
on those, then bugpoint will get confused. To work around this,
add a "bugpoint" mode to multiversioning where it accepts somewhat
corrupted IR and tries to do something reasonable. This mode is not
enabled from julia internally, but is enabled when run through the
external pass interface. Of course the result of a bugpoint reduction
won't be a loadable system image as a result, but bugpoint already
is basically guaranteed to generate a non-loadable image, so that
isn't an issue (since bugpoint is for debugging the compiler itself).