Gate llvm.dbg intrinsic generation behind a flag
Julia is capable of generating llvm.dbg intrinsics. These get
lowered into DWARF debug info and can be used by GDB to find variables.
At least that's the theory. In practice, in our current implementation
they're mostly useless, because we lose location information for local
variables during SSA generation (so we only end up generating these
for arguments anyway, and likely with an incorrect lifetime).
Additionally, by default, Julia does not even register its JIT frames
with GDB, so the information is doubly useless.

Now, generating this information is not super-costly, but does add about
5% to LLVM time in some benchmarks, which is more than I'm willing to pay.

This PR adds a `-gvalues` command line flag (off by default unless building
the sysimage or the GDB listener is manually turned on), which controls
generation of llvm.dbg intrinsics. I suspect nobody would miss them if
we simply deleted them, but I figure maybe somebody wants to try to
improve this situation in the future, so might as well keep it around
as it's off by default.