IterableClosure is an iterator based on a closure
I'm not sure there's already and alternative for this available in Julia, so I came up with this proposal. Please let me know if there's another way available. I'm especially interested in having a way to use do-syntax to create an iterator.

The idea is that defining iterators using the `Base.iterate` is a little bit inconvenient. Then I noticed that it's pretty convenient to create a `Channel`. But there's no equally easy way to create a simpler iterator.

Some languages allow you to create an iterator based on a closure that returns an optional type. The proposed class and its corresponding `iterate` method should allow you to do that. Using this, it's easy to write a function that defines a closure and returns and iterator based on it, as the example illustrates.

```
julia> function myrange(a, b)
           state = a
           IterableClosure() do
               if state <= b
                   prevstate, state = state, state + 1
                   Some(prevstate)
               else
                   nothing
               end
           end
       end
myrange (generic function with 1 method)

julia> for x in myrange(3,5)
           println(x)
       end
3
4
5
```