move ans and err variables to MainInclude module
This hides them from `names()` unfortunately, but means we will not
accidentally discard a variable from the user in Main, either because
the import will fail or the assignment will fail. If we later have
world-versioned bindings, this would also mean you could toggle (between
worlds) between being module-local and a special import value.

This also needed some improvements to docs, so that it would print
docstrings for unassigned globals without also simultaneously claiming
that it did not exist.

Fix #43172
Fix #48299