Avoid unnecessary invalidations for `invoke`d fallbacks
While investigating a report in https://discourse.julialang.org/t/package-load-time-regressions-in-v1-8-beta3/78875
it became clear that some precompilation statements were not effective
due to invalidation.  It turns out that these have the pattern

    f(arg::SpecificType) = invoke(f, Tuple{GenericType}, arg)

When `f` gets compiled for a new subtype of `SpecificType`, the
external cached MethodInstance gets invalidated because of the
reliance on a less specific method.

This is a hack that works around this issue for `copyto!`, but being
able to mark calls as having been issued from an `invoke` would
be a much better solution.